softwaredevelopment/src/main/java/com/software/development/softwaredevelopment/chapter5/BusinessRuleEngine.java
package com.software.development.softwaredevelopment.chapter5;

import java.util.ArrayList;
import java.util.List;

@Deprecated
public class BusinessRuleEngine {

    private final List<Action> actions;
    private final Facts facts;

    public BusinessRuleEngine(Facts facts) {
        this.actions = new ArrayList<>();
        this.facts = facts;
    }

    public void addAction(final Action action) {
        this.actions.add(action);
    }

    public int count() {
        return this.actions.size();
    }

    public void run() {
        this.actions.forEach(action -> action.execute(facts));
    }

}